- Small changes to view layouts
