+++ Gießformkonfigurator.WPF/MVVM/View/Components_BoltView.xaml
@@ -19,7 +19,7 @@
                 <ColumnDefinition Width="auto"/>
                 <ColumnDefinition Width="200"/>
                 <ColumnDefinition Width="50"/>
-                <ColumnDefinition Width="auto"/>
+                <ColumnDefinition Width="152"/>
                 <ColumnDefinition Width="200"/>
             </Grid.ColumnDefinitions>
 
@@ -91,7 +91,7 @@
                 </TextBox.Text>
             </TextBox>
 
-            <TextBlock Text="Außendurchmeser" Grid.Row="3" Grid.Column="0"/>
+            <TextBlock Text="Außendurchmeser " Grid.Row="3" Grid.Column="0"/>
             <TextBox x:Name="OuterDiameter_TextBox" Grid.Row="3" Grid.Column="1">
                 <TextBox.Text>
                     <Binding Path="Bolt.OuterDiameter" Mode="TwoWay" UpdateSourceTrigger="PropertyChanged">
