+++ Gießformkonfigurator.WPF/MVVM/View/Components_CoreView.xaml
@@ -20,7 +20,7 @@
                 <ColumnDefinition Width="auto"/>
                 <ColumnDefinition Width="200"/>
                 <ColumnDefinition Width="50"/>
-                <ColumnDefinition Width="auto"/>
+                <ColumnDefinition Width="152"/>
                 <ColumnDefinition Width="200"/>
             </Grid.ColumnDefinitions>
 
@@ -92,7 +92,7 @@
                 </TextBox.Text>
             </TextBox>
 
-            <TextBlock Text="Außendurchmeser" Grid.Row="3" Grid.Column="0"/>
+            <TextBlock Text="Außendurchmeser " Grid.Row="3" Grid.Column="0"/>
             <TextBox x:Name="OuterDiameter_TextBox" Grid.Row="3" Grid.Column="1">
                 <TextBox.Text>
                     <Binding Path="Core.OuterDiameter" Mode="TwoWay" UpdateSourceTrigger="PropertyChanged">
