--- Gießformkonfigurator.WPF/MVVM/View/Product_MainView.xaml
@@ -1,45 +1,0 @@
-﻿<UserControl x:Class="Giessformkonfigurator.WPF.MVVM.View.Product_MainView"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Giessformkonfigurator.WPF.MVVM.View"
-             mc:Ignorable="d" 
-             d:DesignHeight="900" d:DesignWidth="1600"
-             Background="White">
-    
-    <Border BorderThickness="1" BorderBrush="Black" CornerRadius="4">
-        
-        <Grid>
-            <Grid.RowDefinitions>
-                <RowDefinition Height="100"/>
-                <RowDefinition/>
-            </Grid.RowDefinitions>
-
-            <Border BorderBrush="Gray" BorderThickness="0, 0, 0, 1" Margin="10"/>
-
-            <StackPanel Grid.Row="0" Orientation="Horizontal"  HorizontalAlignment="Center">
-                <RadioButton Content="Produkt Scheibe"
-                             Height="50" Width="200"  
-                             HorizontalAlignment="Center" VerticalAlignment="Center"
-                             FontSize="20" BorderBrush="Black"
-                             Margin="0,0,0,0"
-                             Style="{StaticResource MenuButtonTheme}"
-                             IsChecked="True"
-                             Command="{Binding Product_DiscViewCmd}"/>
-
-                <RadioButton Content="Produkt Cup"
-                             Height="50" Width="200"   
-                             HorizontalAlignment="Center" VerticalAlignment="Center"
-                             FontSize="20"
-                             Style="{StaticResource MenuButtonTheme}"
-                             Command="{Binding Product_CupViewCmd}"/>
-            </StackPanel>
-
-            <ContentControl Grid.Row="1" Grid.Column="1"
-                            Margin="10"
-                            Content="{Binding CurrentView_Product_MainView}"/>
-
-        </Grid>
-    </Border>
-</UserControl>
