--- Gießformkonfigurator.WPF/MVVM/View/Product_MainView.xaml.cs
@@ -1,19 +1,0 @@
-﻿//-----------------------------------------------------------------------
-// <copyright file="Product_MainView.xaml.cs" company="PlaceholderCompany">
-// Copyright (c) PlaceholderCompany. All rights reserved.
-// </copyright>
-//-----------------------------------------------------------------------
-namespace Giessformkonfigurator.WPF.MVVM.View
-{
-    using System.Windows.Controls;
-    /// <summary>
-    /// Interaktionslogik für HinzufuegenView.xaml
-    /// </summary>
-    public partial class Product_MainView : UserControl
-    {
-        public Product_MainView()
-        {
-            InitializeComponent();
-        }
-    }
-}
